Fix missing airdodge/shield tilt coordinates
The horizontal and vertical airdodge coordinates were not defined, and
neither were the 1/2 and 3/4 quadrant airdodge coordinates.
Additionally, the modX/modY modifications were not applied to cardinals
while shielding. They should, however, be the same when holding shield
as they are without shield (as of V3) so that modX can be used for
shield tilt.